Move RecipesBase and RuntimeGeneratedFunctions to extensions
This commit moves RecipesBase and RuntimeGeneratedFunctions from direct
dependencies to weak dependencies with corresponding extensions, reducing
the load time footprint of SciMLBase.

Changes:
- Moved RecipesBase from deps to weakdeps in Project.toml
- Moved RuntimeGeneratedFunctions from deps to weakdeps in Project.toml
- Created SciMLBaseRecipesBaseExt.jl extension containing all @recipe definitions
- Created SciMLBaseRuntimeGeneratedFunctionsExt.jl extension with numargs method
- Removed RecipesBase import from main SciMLBase.jl module
- Removed RuntimeGeneratedFunctions import from main SciMLBase.jl module
- Removed all @recipe function definitions from original source files
- Removed numargs method for RuntimeGeneratedFunctions from utils.jl

The plotting functionality is now only available when RecipesBase is explicitly
loaded, maintaining backward compatibility while reducing the default dependency
footprint. The RuntimeGeneratedFunctions support for numargs is similarly
conditional.
