+++ Project.toml
@@ -22,10 +22,8 @@ Moshi = "2e0e35c7-a2e4-4343-998d-7ef72827ed2d"
 PrecompileTools = "aea7be01-6a6a-4083-8856-8a6e6704d82a"
 Preferences = "21216c6a-2e73-6563-6e65-726566657250"
 Printf = "de0858da-6303-5e67-8744-51eddeeeb8d7"
-RecipesBase = "3cdcf5f2-1ef4-517c-9805-6587b60abb01"
 RecursiveArrayTools = "731186ca-8d62-57ce-b412-fbd966d074cd"
 Reexport = "189a3867-3050-52da-a836-e630ba90ab69"
-RuntimeGeneratedFunctions = "7e49a35a-f44a-4d26-94aa-eba1b4ca6b47"
 SciMLOperators = "c0aeaf25-5076-4817-a8d5-81caf7dfa961"
 SciMLStructures = "53ae85a6-f571-4167-b2af-e1d143709226"
 StaticArraysCore = "1e83bf80-4336-4d27-bf5d-d5a4f845583c"
@@ -41,6 +39,7 @@ PartialFunctions = "570af359-4316-4cb7-8c74-252c00c2016b"
 PyCall = "438e738f-606a-5dbb-bf0a-cddfbfd45ab0"
 PythonCall = "6099a3de-0909-46bc-b1f4-468b9a2dfc0d"
 RCall = "6f49c342-dc21-5d91-9882-a32aef131414"
+RuntimeGeneratedFunctions = "7e49a35a-f44a-4d26-94aa-eba1b4ca6b47"
 Zygote = "e88e6eb3-aa80-5325-afca-941959d7151f"
 
 [extensions]
@@ -51,6 +50,8 @@ SciMLBasePartialFunctionsExt = "PartialFunctions"
 SciMLBasePyCallExt = "PyCall"
 SciMLBasePythonCallExt = "PythonCall"
 SciMLBaseRCallExt = "RCall"
+SciMLBaseRecipesBaseExt = "RecipesBase"
+SciMLBaseRuntimeGeneratedFunctionsExt = "RuntimeGeneratedFunctions"
 SciMLBaseZygoteExt = ["Zygote", "ChainRulesCore"]
 
 [compat]
@@ -81,7 +82,6 @@ Printf = "1.10"
 PyCall = "1.96"
 PythonCall = "0.9.15"
 RCall = "0.14.0"
-RecipesBase = "1.3.4"
 RecursiveArrayTools = "3.35"
 Reexport = "1"
 RuntimeGeneratedFunctions = "0.5.12"
