+++ src/integrator_interface.jl
@@ -773,135 +773,6 @@ end
 
 Base.length(iter::TimeChoiceIterator) = length(iter.ts)
 
-@recipe function f(integrator::DEIntegrator;
-        denseplot = (integrator.opts.calck ||
-                     integrator isa AbstractSDEIntegrator) &&
-                    integrator.iter > 0,
-        plotdensity = 10,
-        plot_analytic = false, vars = nothing, idxs = nothing)
-    if vars !== nothing
-        Base.depwarn(
-            "To maintain consistency with solution indexing, keyword argument vars will be removed in a future version. Please use keyword argument idxs instead.",
-            :f; force = true)
-        (idxs !== nothing) &&
-            error("Simultaneously using keywords vars and idxs is not supported. Please only use idxs.")
-        idxs = vars
-    end
-
-    int_vars = interpret_vars(idxs, integrator.sol)
-
-    if denseplot
-        # Generate the points from the plot from dense function
-        plott = collect(range(integrator.tprev, integrator.t; length = plotdensity))
-        if plot_analytic
-            plot_analytic_timeseries = [integrator.sol.prob.f.analytic(
-                                            integrator.sol.prob.u0,
-                                            integrator.sol.prob.p,
-                                            t) for t in plott]
-        end
-    else
-        plott = nothing
-    end
-
-    dims = length(int_vars[1])
-    for var in int_vars
-        @assert length(var) == dims
-    end
-    # Should check that all have the same dims!
-
-    plot_vecs = []
-    for i in 2:dims
-        push!(plot_vecs, [])
-    end
-
-    labels = String[]# Array{String, 2}(1, length(int_vars)*(1+plot_analytic))
-    strs = String[]
-    varsyms = variable_symbols(integrator)
-    @show plott
-
-    for x in int_vars
-        for j in 2:dims
-            if denseplot
-                if (x[j] isa Integer && x[j] == 0) ||
-                   isequal(x[j], getindepsym_defaultt(integrator))
-                    push!(plot_vecs[j - 1], plott)
-                else
-                    push!(plot_vecs[j - 1], Vector(integrator(plott; idxs = x[j])))
-                end
-            else # just get values
-                if x[j] == 0
-                    push!(plot_vecs[j - 1], integrator.t)
-                elseif x[j] == 1 && !(integrator.u isa AbstractArray)
-                    push!(plot_vecs[j - 1], integrator.u)
-                else
-                    push!(plot_vecs[j - 1], integrator.u[x[j]])
-                end
-            end
-
-            if !isempty(varsyms) && x[j] isa Integer
-                push!(strs, String(getname(varsyms[x[j]])))
-            elseif hasname(x[j])
-                push!(strs, String(getname(x[j])))
-            else
-                push!(strs, "u[$(x[j])]")
-            end
-        end
-        add_labels!(labels, x, dims, integrator.sol, strs)
-    end
-
-    if plot_analytic
-        for x in int_vars
-            for j in 1:dims
-                if denseplot
-                    push!(plot_vecs[j],
-                        u_n(plot_timeseries, x[j], sol, plott, plot_timeseries))
-                else # Just get values
-                    if x[j] == 0
-                        push!(plot_vecs[j], integrator.t)
-                    elseif x[j] == 1 && !(integrator.u isa AbstractArray)
-                        push!(plot_vecs[j],
-                            integrator.sol.prob.f(Val{:analytic}, integrator.t,
-                                integrator.sol[1]))
-                    else
-                        push!(plot_vecs[j],
-                            integrator.sol.prob.f(Val{:analytic}, integrator.t,
-                                integrator.sol[1])[x[j]])
-                    end
-                end
-            end
-            add_labels!(labels, x, dims, integrator.sol, strs)
-        end
-    end
-
-    xflip --> integrator.tdir < 0
-
-    if denseplot
-        seriestype --> :path
-    else
-        seriestype --> :scatter
-    end
-
-    # Special case labels when idxs = (:x,:y,:z) or (:x) or [:x,:y] ...
-    if idxs isa Tuple && (typeof(idxs[1]) == Symbol && typeof(idxs[2]) == Symbol)
-        xlabel --> idxs[1]
-        ylabel --> idxs[2]
-        if length(idxs) > 2
-            zlabel --> idxs[3]
-        end
-    end
-    if getindex.(int_vars, 1) == zeros(length(int_vars)) ||
-       getindex.(int_vars, 2) == zeros(length(int_vars))
-        xlabel --> "t"
-    end
-
-    linewidth --> 3
-    #xtickfont --> font(11)
-    #ytickfont --> font(11)
-    #legendfont --> font(11)
-    #guidefont  --> font(11)
-    label --> reshape(labels, 1, length(labels))
-    (plot_vecs...,)
-end
 
 function step!(integ::DEIntegrator, dt, stop_at_tdt = false)
     (dt * integ.tdir) < 0 * oneunit(dt) && error("Cannot step backward.")
